requirements.txt
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
PyYAML==6.0.2

google-generativeai==0.8.0
langgraph==0.2.0
qdrant-client==1.11.0

# Local AI Models
sentence-transformers==3.0.1
torch>=2.6.0
numpy==1.26.4

# API & Server
fastapi==0.110.0
uvicorn==0.29.0

python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.0

pytest==8.3.0
lxml==5.2.2